moved Bundle, Instructions class from 'model' to 'bean.mojo' package
